% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday_paradox.R
\name{birthday_paradox}
\alias{birthday_paradox}
\alias{birthday_paradox.birthdays}
\title{Calculates any matching birthdays in the data and the probability of having a
matching birthday (i.e. birthday paradox)}
\usage{
birthday_paradox(x, ...)

\method{birthday_paradox}{birthdays}(x, ...)
}
\arguments{
\item{x}{A `birthdays` object}

\item{...}{Arguments to be passed to methods}
}
\value{
A named numeric vector with two elements
}
\description{
Calculates any matching birthdays in the data and the probability of having a
matching birthday (i.e. birthday paradox)
}
\examples{
df <- data.frame(
  name = c("kevin", "stacey"),
  birthday = c(as.Date("2001-01-01"), as.Date("2002-01-02"))
)
birthdays <- birthdays(df)
birthday_paradox(birthdays)
}
