% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthdays.R
\name{birthdays}
\alias{birthdays}
\title{Constructor for creating a subclass of `<data.frame>` called `<birthdays>`}
\usage{
birthdays(x)
}
\arguments{
\item{x}{A `data.frame`}
}
\value{
An `birthdays` object
}
\description{
Constructor for creating a subclass of `<data.frame>` called `<birthdays>`
}
\examples{
df = data.frame(
  name = c("kevin", "stacey"),
  birthday = c(as.Date("2001-01-01"), as.Date("2002-01-02"))
)
birthdays(df)


# generate large randomised `<birthdays>` objects

df <- data.frame(
  name = randomNames::randomNames(10),
  birthday = sample(
    x = seq.Date(
      from = as.Date("2020-01-01"),
      to = as.Date("2022-01-01"),
      by = 1),
    size = 10
  )
)
birthdays <- birthdays(df)
}
