% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthdays_per_month.R
\name{birthdays_per_month}
\alias{birthdays_per_month}
\alias{birthdays_per_month.birthdays}
\title{Count number of birthdays per month in each year from a `<birthdays>` object}
\usage{
birthdays_per_month(x, ...)

\method{birthdays_per_month}{birthdays}(x, ...)
}
\arguments{
\item{x}{An `birthdays` object}

\item{...}{Arguments to be passed on}
}
\value{
A `birthdays` object
}
\description{
Count number of birthdays per month in each year from a `<birthdays>` object
}
\examples{
df = data.frame(
  name = c("kevin", "stacey"),
  birthday = c(as.Date("2001-01-01"), as.Date("2002-01-02"))
)
birthdays <- birthdays(df)
birthdays_per_month(birthdays)
}
