% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-reconstruct.R
\name{birthdays_reconstruct}
\alias{birthdays_reconstruct}
\title{Decides whether `<birthdays>` object can be reconstructed from input}
\usage{
birthdays_reconstruct(x, to)
}
\arguments{
\item{x}{A `data.frame` or subclass of `data.frame` (e.g. `tibble` or
`birthdays`)}

\item{to}{The reference object, in this case an `birthdays` object}
}
\value{
An `birthdays` object (if the input is valid) or a `data.frame`
}
\description{
Uses [`birthdays_can_reconstruct()`] to determine whether the
data input can be reconstructed in a valid `<birthdays>` object. If it can
not, it is returned as a `data.frame`.
}
\keyword{internal}
