% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-reconstruct.R
\name{$<-.birthdays}
\alias{$<-.birthdays}
\title{Set accessor for `<birthdays>` class}
\usage{
\method{$}{birthdays}(x, name) <- value
}
\arguments{
\item{x}{An `birthdays` object}

\item{name}{
    A literal character string or a \link[base]{name} (possibly \link[base]{backtick}
    quoted).  For extraction, this is normally (see under
    \sQuote{Environments}) partially matched to the \code{\link[base]{names}}
    of the object.
  }

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\value{
An `birthdays` object or a `data.frame`
}
\description{
Set accessor for `<birthdays>` class
}
