% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-reconstruct.R
\name{df_reconstruct}
\alias{df_reconstruct}
\title{Transplants the attributes of one input (`to`) to the other input (`x`)}
\usage{
df_reconstruct(x, to)
}
\arguments{
\item{x}{A `data.frame` or subclass of `data.frame` (e.g. `tibble` or
`birthdays`)}

\item{to}{The reference object, in this case an `birthdays` object}
}
\value{
An `birthdays` object
}
\description{
Transplants the attributes of one input (`to`) to the other input (`x`)
}
\keyword{internal}
