% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendDataFrames-package.R
\docType{package}
\name{ExtendDataFrames-package}
\alias{ExtendDataFrames}
\alias{ExtendDataFrames-package}
\title{ExtendDataFrames: Explains how to extend the data frame class in R}
\description{
This package acts as a demonstration on how to extend the data frame class in R, and how to make it compatible with {dplyr} functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joshwlambert/ExtendDataFrames}
  \item \url{https://joshwlambert.github.io/ExtendDataFrames/}
  \item Report bugs at \url{https://github.com/joshwlambert/ExtendDataFrames/issues}
}

}
\author{
\strong{Maintainer}: Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [copyright holder]

}
\keyword{internal}
