% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-reconstruct.R
\name{names<-.birthdays}
\alias{names<-.birthdays}
\title{Set names on `<birthdays>` class}
\usage{
\method{names}{birthdays}(x) <- value
}
\arguments{
\item{x}{an \R object.}

\item{value}{a character vector of up to the same length as \code{x}, or
    \code{NULL}.}
}
\value{
An `<birthdays>` object or a `data.frame`
}
\description{
If the modifying the names invalidates the `<birthdays>` object
the subsetting will return a data frame with the other attributes of the
class preserved.
}
