% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthdays.R
\name{print.birthdays}
\alias{print.birthdays}
\title{Print method for `<birthdays>`}
\usage{
\method{print}{birthdays}(x, ...)
}
\arguments{
\item{x}{An `birthdays` object}

\item{...}{Other arugments to be passed to `print()`}
}
\value{
Invisibly returns input, called for printing side-effects
}
\description{
Print method for `<birthdays>`
}
\examples{
df = data.frame(
  name = c("kevin", "stacey"),
  birthday = c(as.Date("2001-01-01"), as.Date("2002-01-02"))
)
birthdays <- birthdays(df)
print(birthdays)
}
