% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-reconstruct.R
\name{[.birthdays}
\alias{[.birthdays}
\title{Subset function for `<birthdays>` with downgrade-on-subsetting}
\usage{
\method{[}{birthdays}(x, ...)
}
\arguments{
\item{x}{A `birthdays` object}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
Either a `birthdays` object or a `data.frame`
}
\description{
IF the subsetting invalidates the class then a `<data.frame>`
will be returned.
}
