% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthdays.R
\name{validate_birthdays}
\alias{validate_birthdays}
\title{Class validator for `<birthdays>`}
\usage{
validate_birthdays(x)
}
\arguments{
\item{x}{A `birthdays` object}
}
\value{
Returns input invisibly, using for validation side-effects.
}
\description{
Class validator for `<birthdays>`
}
\examples{
df = data.frame(
  name = c("kevin", "stacey"),
  birthday = c(as.Date("2001-01-01"), as.Date("2002-01-02"))
)
birthdays <- birthdays(df)
validate_birthdays(birthdays)
}
